Code cleaning and bug fixes
Cleaned the code and made it more reusable and modular.

Fixed a bug that caused the pixi.js canvas to be attached to the body
instead of the game area.

Game now creates the canvas when the game is started rather than when
page is loaded. Temporarily creates movable bunnies until further game
is implemented