Began the implementation of finishing the build
Preliminary implementation of finishing the build. At the moment when
player or host presses a square in the game area the host is sent an
image of the whole play area and then it is rendered over the host's
play area.

The objects will be rendered on a seperate hidden canvas with
see-through stage and then that canvas will be made into a image to send
to host.

Unfortunately we cannot send objects through Socket.io and thus are
forced to instead send an image file of the objects.