Added a function for adding menu items
pixijs.addToBlockMenu adds a sprite to the block menu which can then be
clicked to add the sprite to play area.
The function has parameters texture and amount, texture specifies
texture to use and amount how many of said elements can be added to the
playArea.
Still needs to add a text which tells the remaining amount