Fixed percentage that doesnt match, added object centering
Percentage shown at the end of the round now should correctly apply to
just the built object, small errors still due to using image height and
width instead of the actual area of the blueprint.

Objects are now centered before sending them to the instructor to show,
this also makes it possible to compare the blueprint to the built
object.